--- src/views/user-settings/FormType.d.ts
@@ -1,7 +1,0 @@
-export interface FormType {
-	username: string | null;
-	display_name: string | null;
-	profile_picture: ArrayBuffer | null;
-	selected_paint: string | null;
-	selected_badge: string | null;
-}
