+++ src/views/user/UserRoot.vue
@@ -1,192 +1,29 @@
 <template>
-	<main class="user-page">
+	<main class="user-root">
 		<template v-if="ctx?.user">
 			<!-- User Details - name tag, roles, channels, etc -->
 			<div class="container">
-				<UserDetails :user="ctx.user" />
+				<UserDetails />
 
-				<router-view class="user-data" />
-				<div ref="userData" class="user-data">
-					<h3 section-title selector="emote-sets">
-						<span> {{ t("user.emote_sets") }}</span>
-
-						<!-- Migrate Button -->
-						<div :style="{ display: 'flex', gap: '0.25em' }">
-							<router-link :to="{ name: 'Migrate' }" class="migrate" :style="{ 'z-index': 1 }">
-								<Button color="accent" fa-icon="check-double" :label="t('migrate.cta')" />
-							</router-link>
-							<Button
-								v-if="actorMayManageSets"
-								:label="t('emote_set.create')"
-								appearance="raised"
-								color="primary"
-								fa-icon="hexagon-plus"
-								selector="emote-set-create"
-								@click="createEmoteSet"
-							/>
-						</div>
-					</h3>
-					<div section-body>
-						<div v-if="ctx.emoteSets.length" selector="emote-set-list">
-							<EmoteSetCard v-for="set of ctx.emoteSets" :key="set.id" :set="set" />
-						</div>
-						<p v-else>{{ t("user.no_sets", [ctx.user?.display_name]) }}</p>
-					</div>
-
-					<!-- Display Channel Emotes -->
-					<UserChannelEmotes v-if="ctx.user" :user="ctx.user" :page-size="cardCount" />
-					<span v-else>{{ t("common.loading") }}...</span>
-
-					<!-- Display Owned Emotes -->
-					<h3 v-if="ctx.user && ctx.ownedEmotes.length" section-title>
-						<span>Owned Emotes ({{ ownedPager.length }})</span>
-						<span selector="search-bar">
-							<TextInput v-model="ownedEmoteSearch" icon="search" :label="t('common.search')" />
-						</span>
-					</h3>
-					<div v-if="ctx.user" section-body>
-						<div class="owned-emotes emote-list">
-							<EmoteCardList :items="pagedOwnedEmotes" />
-						</div>
-						<div v-if="ownedPager.length / cardCount > 1">
-							<Paginator
-								:page="ownedPager.page"
-								:items-per-page="cardCount"
-								:length="ownedPager.length"
-								@change="(change) => (ownedPager.page = change.page)"
-							/>
-						</div>
-					</div>
-
-					<!-- Display Activity -->
-					<h3 v-if="ctx.user && ctx.activity.length" section-title>Activity</h3>
-
-					<div v-if="ctx.user && ctx.emoteSets.length" class="activity-list">
-						<div v-for="log in ctx.activity" :key="log.id">
-							<Activity :target="findActiveSet(log.target_id) ?? ctx.user" :log="log" />
-						</div>
-					</div>
-				</div>
+				<RouterView />
 			</div>
 		</template>
 	</main>
 </template>
 
 <script setup lang="ts">
-import { defineAsyncComponent, reactive, ref, watch, watchEffect } from "vue";
-import { ConvertDecimalToHex } from "@/structures/util/Color";
-import { EmoteSet } from "@/structures/EmoteSet";
-import { ObjectKind } from "@/structures/Common";
-import { useActor } from "@/store/actor";
 import { useContext } from "@/composables/useContext";
-import { useI18n } from "vue-i18n";
-import { useModal } from "@/store/modal";
-import { useObjectSubscription } from "@/composables/useObjectSub";
-import { User } from "@/structures/User";
-import { useSizedRows } from "@/composables/useSizedRows";
-import type { Emote } from "@/structures/Emote";
-import Button from "@/components/utility/Button.vue";
-import EmoteCardList from "../../components/utility/EmoteCardList.vue";
-import EmoteSetCard from "@/components/emote-set/EmoteSetCard.vue";
-import Paginator from "@/views/emote-list/Paginator.vue";
-import TextInput from "@/components/form/TextInput.vue";
-import UserChannelEmotes from "./UserChannelEmotes.vue";
-import UserDetails from "@/views/user/UserDetails.vue";
-
-const ModalCreateEmoteSet = defineAsyncComponent(() => import("@/components/modal/ModalCreateEmoteSet.vue"));
-const Activity = defineAsyncComponent(() => import("@/components/activity/Activity.vue"));
-
-const { t } = useI18n();
+import UserDetails from "./UserDetails.vue";
 
 const ctx = useContext("USER");
-
-const actor = useActor();
-
-const { watchObject } = useObjectSubscription();
-
-const userData = ref<HTMLElement>();
-const cardCount = ref(1);
-const { update: updateSizing } = useSizedRows([128, 160]);
-const sizedRows = ref(0);
-
-const ownedPager = reactive({
-	page: 1,
-	length: 0,
-	start: 0,
-	end: 0,
-});
-const pagedOwnedEmotes = ref([] as Emote[]);
-
-const actorMayManageSets = ref(false);
-
-watchEffect(() => {
-	if (!ctx?.user) return;
-
-	document.documentElement.style.setProperty(
-		"--user-page-sections-color",
-		ctx.user.style && ctx.user.style.color !== 0 ? ConvertDecimalToHex(ctx.user.style.color) : "#FFFFFF40",
-	);
-
-	sizedRows.value = updateSizing(userData.value as HTMLElement).rows;
-	cardCount.value = sizedRows.value * 5;
-	watchObject(ObjectKind.USER, ctx.user);
-
-	actorMayManageSets.value = actor.hasEditorPermission(ctx.user, User.EditorPermission.ManageEmoteSets);
-
-	// Paginate owned emotes
-	ownedPager.length = ctx.ownedEmotes.length;
-	ownedPager.start = (ownedPager.page - 1) * cardCount.value;
-	ownedPager.end = ownedPager.start + cardCount.value;
-	pagedOwnedEmotes.value = ctx.ownedEmotes
-		.filter((e) => ownedEmotesSearched(e.name))
-		.slice(ownedPager.start, ownedPager.end);
-});
-
-// Search
-const ownedEmotesSearched = (s: string) =>
-	!ownedEmoteSearch.value || s.toLowerCase().includes(ownedEmoteSearch.value.toLowerCase());
-
-const ownedEmoteSearch = ref("");
-watch(ownedEmoteSearch, () => {
-	ownedPager.page = 1;
-});
-
-const findActiveSet = (id: string): EmoteSet | null => {
-	for (const set of ctx?.emoteSets ?? []) {
-		if (set.id === id) {
-			return set;
-		}
-	}
-
-	return null;
-};
-
-const modal = useModal();
-const createEmoteSet = () => {
-	if (!ctx?.user) return;
-
-	modal.open("create-emote-set", {
-		component: ModalCreateEmoteSet,
-		props: {
-			user: ctx.user,
-			startingValue: {
-				connections: [null],
-			},
-		},
-		events: {
-			created: (set: EmoteSet) => {
-				ctx.user.emote_sets.push(set);
-			},
-		},
-	});
-};
 </script>
 
 <style lang="scss" scoped>
 @import "@scss/themes.scss";
 
 $smallWidth: 800px;
-.user-page {
+
+main.user-root {
 	.container {
 		display: flex;
 		flex-direction: row;
@@ -198,7 +35,7 @@ $smallWidth: 800px;
 
 		@media screen and (max-width: $smallWidth) {
 			flex-direction: column;
-			.user-data {
+			.user-overview {
 				margin-top: 2em;
 
 				.emote-list {
@@ -206,92 +43,6 @@ $smallWidth: 800px;
 				}
 			}
 		}
-
-		.user-data {
-			display: flex;
-			flex-direction: column;
-			width: 100%;
-
-			@include themify() {
-				background-color: lighten(themed("backgroundColor"), 1);
-			}
-
-			.emote-list {
-				display: flex;
-				flex-wrap: wrap;
-				align-items: flex-start;
-				margin-bottom: 1em;
-				gap: 1em;
-			}
-
-			h3[section-title] {
-				display: flex;
-				justify-content: space-between;
-				align-items: center;
-				height: 3em;
-				font-size: 1.25em;
-				padding-left: 1em;
-				border-bottom-right-radius: 0;
-				border-bottom-left-radius: 0;
-				width: 100%;
-				padding-right: 0.5em;
-
-				&[selector="emote-sets"] {
-					span {
-						z-index: 1;
-					}
-					button {
-						font-size: 1rem;
-					}
-				}
-
-				> [selector="search-bar"] {
-					font-weight: 400;
-					> div {
-						width: min(12em, 25vw);
-					}
-				}
-
-				@include themify() {
-					background-image: linear-gradient(
-						60deg,
-						var(--user-page-sections-color, themed("backgroundColor")) 16em,
-						transparent 0,
-						themed("navBackgroundColor") 0
-					);
-				}
-			}
-			div[section-body] {
-				border-radius: 0.5em;
-				padding: 0.5em;
-				padding-bottom: 1em;
-				@include themify() {
-					background-color: lighten(themed("backgroundColor"), 2);
-				}
-
-				[selector="emote-set-list"] {
-					display: flex;
-					flex-direction: row;
-					flex-wrap: wrap;
-					> .emote-set-card {
-						margin: 0.25em;
-					}
-				}
-			}
-			.section-has-nothing {
-				padding: 1em;
-				@include themify() {
-					color: darken(themed("color"), 20);
-				}
-			}
-		}
 	}
-}
-
-.user-unknown {
-	display: flex;
-	width: 100%;
-	height: 100%;
-	place-content: center;
 }
 </style>
