+++ src/views/user/UserSettingsCosmeticsUpdateModal.vue
